# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet API",
        "description": "Wallet Ledger System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Current user info"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                }
            }
        },
        "/api/wallet/transfer": {
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "Transfer funds to another wallet",
                "operationId": "77b25c3e37e32e898295ad348be12089",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recipient_wallet_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "auto_retry": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer completed"
                    },
                    "202": {
                        "description": "Transfer scheduled for retry"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "Idempotency conflict"
                    },
                    "422": {
                        "description": "Business rejection"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/wallet/transactions/{transaction}/reverse": {
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "Reverse a ledger transaction",
                "operationId": "2fa4a8ae4ec6e1f16fd4ad938c8a3518",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reversal completed"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/wallet": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet for current user",
                "operationId": "e38cf877a4ab9e4edb0e615b86e5caaf",
                "responses": {
                    "200": {
                        "description": "Wallet info"
                    }
                }
            }
        },
        "/api/wallet/deposit": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Deposit funds into wallet",
                "operationId": "3cfaee0be2c751eec777b418717eadce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deposit successful"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "Idempotency conflict"
                    }
                }
            }
        },
        "/api/wallet/withdraw": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Withdraw funds from wallet",
                "operationId": "b00d42f65e945f424f78aa35e2237ac7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Withdrawal successful"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Insufficient funds or validation error"
                    }
                }
            }
        },
        "/api/wallet/statement": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet statement",
                "operationId": "143c32523162f1f9ea6e242bf62a925b",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statement returned"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth"
        },
        {
            "name": "Transfer"
        },
        {
            "name": "Wallet"
        }
    ]
}